prompt-tracker/requirements-monitor.txt
Flask==2.3.3
PyYAML==6.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
boto3==1.28.0